Migrate util tests
