+++ src/tests/lib/sort-semesters.test.ts
@@ -1,20 +1,22 @@
-import {Semester} from '.prisma/client';
-import sortSemesters from './sort-semesters';
+import {Semester} from '@prisma/client';
+import test from 'ava';
+import sortSemesters from '~/lib/sort-semesters';
 
-describe('sortSemesters', () => {
-	it('should sort correctly', () => {
-		expect(sortSemesters([
+test('sorts semesters', t => {
+	t.deepEqual(
+		sortSemesters([
 			{year: 2022, semester: Semester.FALL},
 			{year: 2021, semester: Semester.FALL},
 			{year: 2021, semester: Semester.SUMMER},
 			{year: 2021, semester: Semester.SPRING},
 			{year: 2020, semester: Semester.FALL}
-		])).toEqual([
+		]),
+		[
 			{year: 2020, semester: Semester.FALL},
 			{year: 2021, semester: Semester.SPRING},
 			{year: 2021, semester: Semester.SUMMER},
 			{year: 2021, semester: Semester.FALL},
 			{year: 2022, semester: Semester.FALL}
-		]);
-	});
+		]
+	);
 });
